error handling

+++ webUI/openqa_web.sh
@@ -3,8 +3,8 @@ set -e
 REPOSITORY=openqa_webui
 
 usage() {
-		echo "Usage: $0 [-b|-c|-h][-d <openQA_path>]"
-		echo "To start the webUI container, just run this script without any options."
+		echo -e "\nUsage: $0 [-b|-c|-h][-d <openQA_debug_path>]"
+		echo -e "\nRun the webUI container with './openqa_web.sh'"
 		echo "Options:"
 		echo "	-b	Build the web UI container image."
 		echo "	-c	Get and run createhdds to provide images unavailable through fedoraproject.org."
@@ -20,7 +20,7 @@ get_openqa_webui_image_id() {
 build_openqa_webui_image() {
 	CHANGES=$(git diff --name-only HEAD -- | grep -Ev "(client.conf|openqa.ini)$") || true
 	if [ -n "$CHANGES" ]; then
-		echo "Stash or commit changes before building image. Exiting."
+		echo "Stash or commit changes before building image."
 		exit 1
 	fi
 
@@ -48,16 +48,24 @@ run_createhdds() {
 	fi
 
 	cd createhdds
+	# createhdds must run in the background so that it can be interrupted by this shell script
 	./createhdds.py all -c > temp.log 2>&1 &
+
+	# Exit if there are any errors e.g. missing dependencies preventing createhdds from running.
+	sleep 1
+	if grep -q "Traceback" temp.log; then
+		cat temp.log
+		exit 1;
+	fi
 	PID=$!
 
-	# while createhdds is running, check if it is hanging and if so, then
-	# copy the temp file, force it to stop, and then start it again
+	# While createhdds is running, check if it is hanging and if so, then
+	# copy the temp file, force createhdds to stop, and then restart it
 	while kill -0 $PID 2>/dev/null; do
 		FILE_NAME=$(grep "qcow2" temp.log | head -n 1 | sed -n 's/.*\(disk_[^ ]*.qcow2\).*/\1/p')
 		sleep 5
 		if [[ -f "$FILE_NAME.tmp" ]]; then
-			echo "copying $FILE_NAME.tmp"
+			echo "-->copying $FILE_NAME.tmp"
 			cp "$FILE_NAME.tmp" $FILE_NAME
 			kill -SIGINT $PID
 			rm temp.log
@@ -79,6 +87,11 @@ run_createhdds() {
 	cd ..
 }
 
+# For all options, make sure that this script is running from inside the webUI directory.
+# The webUI directory is the context directory for the Dockerfile and is also
+# where the container expects to find assets and files that will be bound in to the container.
+cd $(dirname "$0")
+
 # Parse options
 while getopts ":bchd:" opt; do
 	case ${opt} in
@@ -94,7 +107,15 @@ while getopts ":bchd:" opt; do
 			usage
 			;;
 		d )
-			openQA_path=$OPTARG
+			openQA_debug_path=$OPTARG
+			if [[ $openQA_debug_path == */ ]]; then
+				openQA_debug_path=${openQA_debug_path%/}
+			fi
+
+			debug_arg="-v $openQA_debug_path/script/:/usr/share/openqa/script/:z \
+				-v $openQA_debug_path/lib/OpenQA/:/usr/share/openqa/lib/OpenQA/:z \
+				-v $openQA_debug_path/assets/:/usr/share/openqa/assets/:z \
+				-v $openQA_debug_path/templates/:/usr/share/openqa/templates/:z "
 			;;
 		\? )
 			echo "Invalid option: $OPTARG" 1>&2
@@ -113,21 +134,20 @@ if [ $(($# - ${OPTIND:-0} + 1)) -gt 0 ]; then
 	exit 1
 fi
 
-if [[ -n "$openQA_path" ]]; then
-	if [[ $openQA_path == */ ]]; then
-			openQA_path=${openQA_path%/}
-	fi
-
-	debug_arg="-v $openQA_path/script/:/usr/share/openqa/script/:z \
-		-v $openQA_path/lib/OpenQA/:/usr/share/openqa/lib/OpenQA/:z \
-		-v $openQA_path/assets/:/usr/share/openqa/assets/:z \
-		-v $openQA_path/templates/:/usr/share/openqa/templates/:z "
+get_openqa_webui_image_id
+if [ -z "$OPENQA_WEBUI_IMAGE_ID" ]; then
+	echo "There is no 'openqa_webui' container image available."
+	echo "Run 'openqa_web.sh -b' build the image."
+	exit 1
 fi
 
 if [ ! -d "$PWD/hdd" ] && [ ! -L "$PWD/hdd" ]; then
 	mkdir "$PWD/hdd"
-	run_createhdds
+	echo -e "\nCreating new hdd asset directory."
+	echo "--> Warning: Some tests need images that must be built locally with createhdds."
+	echo -e "--> Run 'openqa_web.sh -c' to get and run createhdds.\n"
 fi
+
 if [ ! -d "$PWD/iso" ] && [ ! -L "$PWD/iso" ]; then
 	mkdir "$PWD/iso"
 	# TODO find a way to get any changes to cloudinit. E.g.
@@ -141,12 +161,6 @@ if [ ! -d "$PWD/tests" ] && [ ! -L "$PWD/tests" ]; then
 	mkdir "$PWD/tests"
 fi
 
-get_openqa_webui_image_id
-if [ -z "$OPENQA_WEBUI_IMAGE_ID" ]; then
-	echo "Building 'openqa_webui' image."
-	build_openqa_webui_image
-	get_openqa_webui_image_id
-fi
 
 podman run -p 8080:80 \
 -v $PWD/hdd:/var/lib/openqa/share/factory/hdd:z \
